backendinator/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
certifi==2024.8.30
charset-normalizer==3.4.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
coloredlogs==15.0.1
environs==9.5.0
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.3.25
fsspec==2024.10.0
ftfy==6.3.1
grpcio==1.68.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
Jinja2==3.1.3
MarkupSafe==3.0.2
marshmallow==3.23.1
milvus-lite==2.4.10
milvus-model==0.2.8
mpmath==1.3.0
networkx==3.2.1
numpy==2.1.3
onnxruntime==1.20.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.3
pymilvus==2.4.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scipy==1.14.1
setuptools==75.5.0
six==1.16.0
sympy==1.13.1
tokenizers==0.20.3
torch==2.5.1
torchvision==0.2.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
wcwidth==0.2.13
